requirements.txt
pbr>=0.6,!=0.7,<1.0
Django>=1.5
python-novaclient
python-heatclient
python-keystoneclient
pyaml
